requirements.txt
asgiref==3.2.6
azure-common==1.1.25
azure-core==1.3.0
azure-nspkg==3.0.2
azure-storage-blob==2.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-storages==1.9.1
gunicorn==20.0.4
idna==2.9
isodate==0.6.0
msrest==0.6.11
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
